requirements.txt
python-dotenv
requests
selenium
undetected-chromedriver